Extract getFilterOptions into a function
